authenfication

+++ src/main/java/com/gestion/gestionUsers/controller/RoleController.java
@@ -6,6 +6,7 @@ package com.gestion.gestionUsers.controller;
 
 import com.gestion.gestionUsers.model.Role;
 import com.gestion.gestionUsers.repository.RoleRepository;
+import com.gestion.gestionUsers.repository.UserRepository;
 import javax.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -25,36 +26,73 @@ public class RoleController {
     
     private RoleRepository roleRepository;
     
+    @Autowired
+    private UserRepository userRepository;
+    
     @GetMapping("/listroles")
     public String ShowListRole(Model model){
         model.addAttribute("roles", roleRepository.findAll());
-        return "layouts/roles";
+        return "pages/tables/roles";
     
     }
     
     @GetMapping("/createRole")
     public String FormCreateRole(Model model){
         model.addAttribute("role", new Role());
-        return "layouts/addrole";
+        return "pages/forms/createrole";
     }
 
 
     @PostMapping("/createRole")
     public String addUser(@Valid Role role, BindingResult result, Model model){
-        if (result.hasErrors()){
-            return "layouts/addrole";
+        try {
+            if (result.hasErrors()) {
+                model.addAttribute("errors", result.getAllErrors());
+                return "/pages/forms/createrole";
+                } else{
+                    if (roleRepository.existsByName(role.getName())) {
+                        model.addAttribute("message", role.getName());
+                        return "/pages/forms/createrole";
+                    }
+
+                    roleRepository.save(role);
+                    return "redirect:/listroles";
+                    }
+        } catch (Exception e) {
+           
+            e.getMessage();
         }
-        
-        roleRepository.save(role);
-        return "redirect:/listroles";
+        return null;
     }
+
+    @GetMapping("/editrole/{id}")
+    public String showUpdateForm(@PathVariable("id") long id, Model model) {
+        Role role = roleRepository.findById(id)
+        .orElseThrow(() -> new IllegalArgumentException("Invalid user Id:" + id));
+        
+        model.addAttribute("role", role);
+        //model.addAttribute("roles", roleRepository.findAll());
+        return "pages/forms/editrole";
+}
 
     @GetMapping("/deleterole/{id}")
     public String deleteRole(@PathVariable("id") long id, Model model) {
     Role role = roleRepository.findById(id)
       .orElseThrow(() -> new IllegalArgumentException("Invalid user Id:" + id));
-    roleRepository.delete(role);
-    return "redirect:/listroles";
+
+    // Vérifier si le rôle est assigné à des utilisateurs
+    if (!userRepository.existsByRolesContains(role)) {
+        // Aucun utilisateur n'a ce rôle, vous pouvez le supprimer
+        roleRepository.delete(role);
+        return "redirect:/listroles";
+
+    } else {
+        // Le rôle est assigné à des utilisateurs, afficher un message d'erreur
+        model.addAttribute("error", "Cannot delete role. It is assigned to one or more users.");
+        return "pages/samples/error";  
+    }
+      
+    // roleRepository.delete(role);
 }
 
 }
